Update return values from numericValues:fromFilename from [Double] to [Int]

+++ IIQM/src/TextReader.swift
@@ -10,7 +10,7 @@ import Foundation
 
 ///Used to simulate asynchronus data from a text file
 class TextReader {
-    func numericValues(from filename: String) throws -> [Double] {
+    func numericValues(from filename: String) throws -> [Int] {
         do {
             //locate the file in the app bundle
             guard let url = Bundle.main.url(forResource: filename, withExtension: "txt")
@@ -20,16 +20,16 @@ class TextReader {
             let contents = try String(contentsOf: url, encoding: .ascii).trimmingCharacters(in: .whitespacesAndNewlines)
             
             //convert the file from a String to an array of Int
-            let lines = contents.components(separatedBy: .newlines).map { Double($0)! }
+            let lines = contents.components(separatedBy: .newlines).map { Int($0)! }
             
             return lines
             
         } catch FileError.invalidFileName {
             print("Unable to read file: \(filename).txt")
-            return [Double]()
+            return [Int]()
         } catch {
             print(error)
-            return [Double]()
+            return [Int]()
         }
         
     }
